requirements.txt
git+https://github.com/levany02/docx2md.git
google-cloud-storage
striprtf